+++ scrapers/news_scraper/newspaper_config/animal_politico.py
