+++ scrapers/news_scraper/newspaper_config/proceso.py
